Changes to VectorR2 int casting * Added methods to cast VectorR2 x and y to int (solves limit problems); * Used said methods to cast VectorR2 coordinates to int in code; * Fixed a VectorR2 creation where x value was cast to int, not double;

+++ src/State.cpp
@@ -227,7 +227,7 @@ VectorR2 State::randomCoordinatesWithMagnitude(
 ) const noexcept {
   double random_angle = M_PI * (rand() % 2000)/1000.0;
 
-  return VectorR2((int) coordinates_magnitude, 0)\
+  return VectorR2((double) coordinates_magnitude, 0)\
     .clockwiseRotatedVector(random_angle);
 };
